+++ CHANGELOG.md
@@ -34,6 +34,10 @@ All notable changes to the "rmarkdown" extension will be documented in this file
 
 ## [0.0.6] - 2020-04-14
 
+### Added
+
+- 'Blogdown: Serve Site' command
+
 ### Changed
 
 - Slugify function preserves CJK characters